LargeScaleIndexer/SourceCode/requirements.txt
Flask
torch
torchvision
requests
Flask-Cors
tensorflow
librosa
python-dotenv